file name change, missed a uint conversion

+++ MultiSig/MultiSigWalletModded.sol
@@ -216,7 +216,7 @@ contract MultiSigWallet {
     /// @param value Transaction ether value.
     /// @param data Transaction data payload.
     /// @return Returns transaction ID.
-    function submitTransaction(address destination, uint value, bytes data)
+    function submitTransaction(address destination, uint256 value, bytes data)
         public
         returns (uint256 transactionId)
     {
